+++ packages/docs/src/constants/apiTableHeaders.ts
@@ -43,12 +43,6 @@ export const API_TABLE_HEADERS: Record<HeaderList, string[]> = {
 		'type',
 		'default',
 		'description'
-	],
-	errorMessages: [
-		'name',
-		'type',
-		'default',
-		'description'
 	]
 };
