+++ packages/docs/src/content/regles-validation/required.md
@@ -3,13 +3,11 @@ title: Règle de validation required
 description: La règle `required` permet de rendre un champ de formulaire obligatoire.
 ---
 
-## Utilisation
+<doc-tabs>
 
-<doc-indent>
+<doc-tab-item label="Utilisation">
 
 Pour utiliser la règle de validation `required`, vous devez l'importer :
-
-</doc-indent>
 
 ```ts
 import { required } from '@cnamts/vue-dot/src/rules/required';
@@ -19,26 +17,30 @@ Puis l'ajouter dans un tableau qui sera utilisé sur un champ de formulaire via 
 
 ```ts
 rules = [
-    required
+	required
 ];
 ```
 
-## Modifier les messages d'erreur
+## Messages d'erreur
 
-<doc-indent>
-
-Il est possible de modifier les messages d’erreur par défaut, pour cela vous devez importer la fonction `requiredFn` :
-
-</doc-indent>
+Vous pouvez modifier les messages d’erreur par défaut en important la fonction `requiredFn` :
 
 ```ts
 import { requiredFn } from '@cnamts/vue-dot/src/rules/required';
 ```
 
-Ensuite, vous devez lui passer en argument un objet qui contient les messages d’erreur utilisés par cette règle de validation :
+Puis en passant en argument un objet contenant les messages d’erreur utilisés par cette règle de validation :
 
 ```ts
 required = requiredFn({
-    default: 'erreur'
+	default: 'Ce champ est requis pour valider le formulaire.'
 });
 ```
+
+</doc-tab-item>
+
+<doc-tab-item label="API">
+<doc-api name="rules/required"></doc-api>
+</doc-tab-item>
+
+</doc-tabs>
