+++ packages/docs/src/data/api/rules/required.ts
@@ -1,0 +1,12 @@
+import { Api } from '~/types';
+import { ruleMessages } from '../shared/ruleMessages';
+
+export const api: Api = {
+	required: {
+		arguments: [
+			...ruleMessages(`{
+	default: 'Le champ est requis.'
+}`)
+		]
+	}
+};
