+++ packages/docs/src/data/api/shared/mixins/widthable.ts
