+++ packages/docs/src/data/api/shared/ruleMessages.ts
@@ -1,0 +1,13 @@
+import { Props } from '~/types';
+
+export function ruleMessages(example?: string): Props {
+	return [
+		{
+			name: 'errorMessages',
+			type: 'ErrorMessages',
+			description: 'Les messages affichés lorsque la règle n’est pas respectée.',
+			default: 'defaultErrorMessages',
+			example
+		}
+	];
+}
