+++ packages/docs/src/data/api/table-toolbar.ts
@@ -1,6 +1,6 @@
 import { Api } from '~/types';
 
-import { customizable } from './mixins/customizable';
+import { customizable } from './shared/mixins/customizable';
 
 export const api: Api = {
 	TableToolbar: {
