+++ packages/docs/src/types/index.d.ts
@@ -19,7 +19,7 @@ type ItemSlot = Pick<ApiProp, 'name' | 'description' | 'props'>;
 type ItemEvent = Pick<ApiProp, 'name' | 'description' | 'value'>;
 type ItemFunction = Pick<ApiProp, 'name' | 'signature' | 'description'>;
 type ItemModifier = Pick<ApiProp, 'name' | 'type' | 'description' | 'snippet'>;
-type ItemArgument = Pick<ApiProp, 'name' | 'type' | 'default' | 'description' | 'example' | 'snippet'>;
+type ItemArgument = Pick<ApiProp, 'name' | 'type' | 'default' | 'required' | 'description' | 'example' | 'snippet'>;
 type ErrorMessages = Pick<ApiProp, 'name' | 'type' | 'default' | 'description'>;
 
 export type Props = ItemValue[];
